requirements.txt
tk
pymupdf
pandas
openpyxl
pillow
pyinstaller
